Share RPC param table between client and server
This helps keep the client and server in sync, and avoids duplication of logic. There is a single
table of params. For each operation, it contains whether each param is a string or some other JSON
value and whether it’s optional or not. The client uses all of this info, but the server references
it to determine if the right number of arguments are provided.

NB: This leaves the (now-redundant) server-side parameter range checks in place, as it will reduce
conflicts if/when we backport any changes related to keyed arguments.
